refine Event Controller

+++ controllers/EventController.py
@@ -1,5 +1,8 @@
+from datetime import datetime
 from flask import request, jsonify
-from markupsafe import escape
+from expetions.data import NotFound
+from expetions.token import InvlidToken, NoToken
+from utils.get_current_user import get_current_user
 from utils.json_encode import json_encode
 from bson.objectid import ObjectId
 
@@ -8,19 +11,40 @@ class EventController:
   @staticmethod
   def get_events(db):
     try:
-      events = db.events.find()
-      response = []
+      user = get_current_user(db)
+      events = db.events.find({'userId': user['_id']})
+      formattedEvents = [json_encode(event, ['userId', 'categoryId']) for event in events]
     
-      return jsonify([json_encode(event) for event in events])
+      return jsonify(refactor_get_events_response(db, formattedEvents))
 
+    except NoToken:
+      return {'message': 'No Token Provided'}, 400
+    except InvlidToken:
+      return {'message': 'Invalid Token'}, 400
     except:
       return {'message': 'An error occurred'}, 500
 
   @staticmethod
   def get_one_event(db, id):
     try:
-      event = db.events.find_one({'_id': ObjectId(id)})
-      return jsonify(json_encode(event))
+      user = get_current_user(db)
+
+      event = db.events.find_one({'_id': ObjectId(id), 'userId': user['_id']})
+      if event is None:
+        raise NotFound
+      
+      category = db.categories.find_one({'_id': event['categoryId']})
+
+      formattedEvent = json_encode(event, ['userId', 'categoryId'])
+      formattedEvent['category'] = json_encode(category, ['userId'])
+
+      return jsonify(formattedEvent)
+    except NoToken:
+      return {'message': 'No Token Provided'}, 400
+    except InvlidToken:
+      return {'message': 'Invalid Token'}, 400
+    except NotFound:
+      return {'message': 'Event not found'}, 404
     except:
       return {'message': 'An error occurred'}, 500
     
@@ -28,20 +52,35 @@ class EventController:
   @staticmethod
   def insert_event(db):
     try:
+      user = get_current_user(db)
+
+      category = db.categories.find_one({'_id': ObjectId(request.form['categoryId']), 'userId': user['_id']})
+      if category is None:
+        return {'message': 'Category not found'}
+
       created_event = db.events.insert_one({
-        'userId':request.form['userId'],
-        'categoryId':request.form['categoryId'],
+        'userId': user['_id'],
+        'categoryId': category['_id'],
         'title':request.form['title'],
         'description':request.form['description'],
         'type':request.form['type'],
-        'dateStart':request.form['dateStart'],
-        'duration':request.form['duration'],
-        'isAllDay':request.form['isAllDay'],
+        'dateStart': datetime.strptime(request.form['dateStart'], '%Y-%m-%d %I:%M%z'),
+        'duration': int(request.form['duration']),
+        'isAllDay':True if request.form['isAllDay'] == 'true' else False,
       })
       event = db.events.find_one({'_id': created_event.inserted_id})
 
-      return jsonify(json_encode(event))
+      formattedEvent = json_encode(event, ['userId', 'categoryId'])
+      formattedEvent['category'] = json_encode(category, ['userId'])
 
+      return jsonify(formattedEvent)
+
+    except NoToken:
+      return {'message': 'No Token Provided'}, 400
+    except InvlidToken:
+      return {'message': 'Invalid Token'}, 400
+    except NotFound:
+      return {'message': 'Event not found'}, 404
     except:
       return {'message': 'An error occurred'}, 500
 
@@ -49,33 +88,73 @@ class EventController:
   @staticmethod
   def update_event(db, id):
     try:
-      updated_event = db.events.update_one({'_id': ObjectId(id)}, {
+      user = get_current_user(db)
+
+      event = db.events.find_one({'_id': ObjectId(id), 'userId': user['_id']})
+      if event is None:
+        raise NotFound
+
+      category = db.categories.find_one({'_id': ObjectId(request.form['categoryId']), 'userId': user['_id']})
+      if category is None:
+        return {'message': 'Category not found'}
+
+      db.events.update_one({'_id': ObjectId(id)}, {
         '$set': {
-          'userId':request.form['userId'],
-          'categoryId':request.form['categoryId'],
+          'userId': user['_id'],
+          'categoryId':category['_id'],
           'title':request.form['title'],
           'description':request.form['description'],
           'type':request.form['type'],
-          'dateStart':request.form['dateStart'],
-          'duration':request.form['duration'],
-          'isAllDay':request.form['isAllDay'],
+          'dateStart':datetime.strptime(request.form['dateStart'], '%Y-%m-%d %I:%M%z'),
+          'duration':int(request.form['duration']),
+          'isAllDay':True if request.form['isAllDay'] == 'true' else False,
         }
       })
+
       event = db.events.find_one({'_id': ObjectId(id)})
+      
+      formattedEvent = json_encode(event, ['userId', 'categoryId'])
+      formattedEvent['category'] = json_encode(category, ['userId'])
 
-      return jsonify(json_encode(event))
+      return jsonify(formattedEvent)
 
+    except NoToken:
+      return {'message': 'No Token Provided'}, 400
+    except InvlidToken:
+      return {'message': 'Invalid Token'}, 400
+    except NotFound:
+      return {'message': 'Event not found'}, 404
     except:
       return {'message': 'An error occurred'}, 500
   
   @staticmethod
   def delete_event(db, id):
     try:
-      updated_event = db.events.delete_one({'_id': ObjectId(id)})
+      user = get_current_user(db)
+
+      event = db.events.find_one({'_id': ObjectId(id), 'userId': user['_id']})
+      if event is None:
+        raise NotFound
+
+      db.events.delete_one({'_id': ObjectId(id)})
 
       return {
         'message': 'Deleted successfully'
       }
-    
+    except NoToken:
+      return {'message': 'No Token Provided'}, 400
+    except InvlidToken:
+      return {'message': 'Invalid Token'}, 400
+    except NotFound:
+      return {'message': 'Event not found'}, 404
     except:
       return {'message': 'An error occurred'}, 500
+
+def refactor_get_events_response(db, events):
+    result = []
+    for event in events:
+      category = db.categories.find_one({'_id': ObjectId(event['categoryId'])})
+      event['category'] = json_encode(category, ['userId'])
+      result.append(event)
+
+    return result
